+++ Plugin/GoalOrder.php
@@ -14,6 +14,7 @@ use Magento\Checkout\Model\Session;
 use Magento\Checkout\Model\Session\SuccessValidator;
 use Magento\Framework\Controller\ResultInterface;
 use Magento\Framework\View\Result\Page;
+use Magento\Sales\Api\Data\OrderInterface;
 use PixelOpen\Plausible\Helper\Config;
 use PixelOpen\Plausible\Session\Goals;
 
@@ -25,38 +26,47 @@ class GoalOrder
 
     protected SuccessValidator $successValidator;
 
+    protected Config $config;
+
     /**
      * @param Goals $goals
      * @param Session $checkoutSession
      * @param SuccessValidator $successValidator
+     * @param Config $config
      */
     public function __construct(
         Goals $goals,
         Session $checkoutSession,
-        SuccessValidator $successValidator
+        SuccessValidator $successValidator,
+        Config $config
     ) {
         $this->goals = $goals;
         $this->checkoutSession = $checkoutSession;
         $this->successValidator = $successValidator;
+        $this->config = $config;
     }
 
     /**
      * Add goal when order succeeds
      *
      * @param Success $subject
-     * @return array
+     * @return null
      */
-    public function beforeExecute(Success $subject): array
+    public function beforeExecute(Success $subject)
     {
         if (!$this->successValidator->isValid()) {
-            return [];
+            return null;
         }
 
         $order = $subject->getOnepage()->getCheckout()->getLastRealOrder();
 
-        $this->goals->add(Config::PLAUSIBLE_GOAL_ORDER, ['total' => (float)$order->getGrandTotal()]);
+        if (!$order) {
+            return null;
+        }
 
-        return [];
+        $this->goals->add(Config::PLAUSIBLE_GOAL_ORDER, [], $this->getRevenue($order));
+
+        return null;
     }
 
     /**
@@ -73,5 +83,23 @@ class GoalOrder
         }
 
         return $result;
+    }
+
+    /**
+     * @param OrderInterface $order
+     * @return mixed[]
+     */
+    public function getRevenue(OrderInterface $order): array
+    {
+        if (!$this->config->isRevenueTrackingEnabled()) {
+            return [];
+        }
+
+        return [
+            'revenue' => [
+                'currency' => $order->getOrderCurrencyCode(),
+                'amount' => (float)$order->getGrandTotal(),
+            ],
+        ];
     }
 }